testing/docker: set default value for CTNG_UID/CTNG_GID
Make the creation of docker images easier so that CTNG_UID/CTNG_GID have
a default value if it's not explicitly specified when building. This
will allow publishing of images on various package repositories (e.g.
docker hub, gitlab containers). dmgr.sh can still be used to set the
UID/GID to that of the current user when building a custom container.

Signed-off-by: Chris Packham <judge.packham@gmail.com>