CI: Create continuous-integration-workflow.yml
Create a continuous integration workflow that builds a few sample
configurations.

Future improvements would be to have a single job that builds ct-ng and
shares the artifacts with the matrix jobs for building the individual
toolchains. It would also be a good idea to fetch and cache the various
source tarballs.

Signed-off-by: Chris Packham <judge.packham@gmail.com>